+++ govcd/openapi_test.go
@@ -8,7 +8,6 @@ package govcd
 
 import (
 	"encoding/json"
-	"fmt"
 	"net/http"
 	"net/url"
 	"regexp"
@@ -23,13 +22,16 @@ import (
 // Test_OpenApiRawJsonAuditTrail uses low level GET function to test out that pagination really works. It is an example
 // how to fetch response from multiple pages in RAW json messages without having defined as struct.
 func (vcd *TestVCD) Test_OpenApiRawJsonAuditTrail(check *C) {
-	minimumRequiredApiVersion := "33.0"
-	skipOpenApiEndpointTest(vcd, check, "1.0.0/auditTrail", minimumRequiredApiVersion)
-	urlRef, err := vcd.client.Client.OpenApiBuildEndpoint("1.0.0/auditTrail")
+	endpoint := types.OpenApiPathVersion1_0_0 + types.OpenApiEndpointAuditTrail
+	skipOpenApiEndpointTest(vcd, check, endpoint)
+	apiVersion, err := vcd.client.Client.checkOpenApiEndpointCompatibility(endpoint)
+	check.Assert(err, IsNil)
+
+	urlRef, err := vcd.client.Client.OpenApiBuildEndpoint(endpoint)
 	check.Assert(err, IsNil)
 
 	// Get a timestamp after which endpoint contains at least 10 elements
-	filterTimeStamp := getAuditTrailTimestampWithElements(10, check, vcd, minimumRequiredApiVersion, urlRef)
+	filterTimeStamp := getAuditTrailTimestampWithElements(10, check, vcd, apiVersion, urlRef)
 
 	// Limit search of audits trails to the last 12 hours so that it doesn't take too long and set pageSize to be 1 result
 	// to force following pages
@@ -39,7 +41,7 @@ func (vcd *TestVCD) Test_OpenApiRawJsonAuditTrail(check *C) {
 	queryParams.Add("sortDesc", "timestamp")
 
 	allResponses := []json.RawMessage{{}}
-	err = vcd.vdc.client.OpenApiGetAllItems(minimumRequiredApiVersion, urlRef, queryParams, &allResponses)
+	err = vcd.vdc.client.OpenApiGetAllItems(apiVersion, urlRef, queryParams, &allResponses)
 
 	check.Assert(err, IsNil)
 	check.Assert(len(allResponses) > 1, Equals, true)
@@ -56,10 +58,12 @@ func (vcd *TestVCD) Test_OpenApiRawJsonAuditTrail(check *C) {
 // Test_OpenApiInlineStructAuditTrail uses low level GET function to test out that get function can unmarshal directly
 // to user defined inline type
 func (vcd *TestVCD) Test_OpenApiInlineStructAuditTrail(check *C) {
-	minimumRequiredApiVersion := "33.0"
-	skipOpenApiEndpointTest(vcd, check, "1.0.0/auditTrail", minimumRequiredApiVersion)
+	endpoint := types.OpenApiPathVersion1_0_0 + types.OpenApiEndpointAuditTrail
+	skipOpenApiEndpointTest(vcd, check, endpoint)
+	apiVersion, err := vcd.client.Client.checkOpenApiEndpointCompatibility(endpoint)
+	check.Assert(err, IsNil)
 
-	urlRef, err := vcd.client.Client.OpenApiBuildEndpoint("1.0.0/auditTrail")
+	urlRef, err := vcd.client.Client.OpenApiBuildEndpoint(endpoint)
 	check.Assert(err, IsNil)
 
 	// Inline type
@@ -101,7 +105,7 @@ func (vcd *TestVCD) Test_OpenApiInlineStructAuditTrail(check *C) {
 	filterTime := time.Now().Add(-6 * time.Hour).Format(types.FiqlQueryTimestampFormat)
 	queryParams.Add("filter", "timestamp=gt="+filterTime)
 
-	err = vcd.vdc.client.OpenApiGetAllItems(minimumRequiredApiVersion, urlRef, queryParams, &allResponses)
+	err = vcd.vdc.client.OpenApiGetAllItems(apiVersion, urlRef, queryParams, &allResponses)
 
 	check.Assert(err, IsNil)
 	check.Assert(len(allResponses) > 1, Equals, true)
@@ -131,9 +135,9 @@ func (vcd *TestVCD) Test_OpenApiInlineStructAuditTrail(check *C) {
 // 9. Delete role once again
 func (vcd *TestVCD) Test_OpenApiInlineStructCRUDRoles(check *C) {
 	endpoint := types.OpenApiPathVersion1_0_0 + types.OpenApiEndpointRoles
-	minimumRequiredApiVersion, err := vcd.client.Client.checkOpenApiEndpointCompatibility(endpoint)
+	apiVersion, err := vcd.client.Client.checkOpenApiEndpointCompatibility(endpoint)
 	check.Assert(err, IsNil)
-	skipOpenApiEndpointTest(vcd, check, endpoint, minimumRequiredApiVersion)
+	skipOpenApiEndpointTest(vcd, check, endpoint)
 
 	// Step 1 - Get all roles
 	urlRef, err := vcd.client.Client.OpenApiBuildEndpoint(endpoint)
@@ -148,7 +152,7 @@ func (vcd *TestVCD) Test_OpenApiInlineStructCRUDRoles(check *C) {
 	}
 
 	allExistingRoles := []*InlineRoles{{}}
-	err = vcd.vdc.client.OpenApiGetAllItems(minimumRequiredApiVersion, urlRef, nil, &allExistingRoles)
+	err = vcd.vdc.client.OpenApiGetAllItems(apiVersion, urlRef, nil, &allExistingRoles)
 	check.Assert(err, IsNil)
 
 	// Step 2 - Get all roles using query filters
@@ -162,7 +166,7 @@ func (vcd *TestVCD) Test_OpenApiInlineStructCRUDRoles(check *C) {
 
 		expectOneRoleResultById := []*InlineRoles{{}}
 
-		err = vcd.vdc.client.OpenApiGetAllItems(minimumRequiredApiVersion, urlRef2, queryParams, &expectOneRoleResultById)
+		err = vcd.vdc.client.OpenApiGetAllItems(apiVersion, urlRef2, queryParams, &expectOneRoleResultById)
 		check.Assert(err, IsNil)
 		check.Assert(len(expectOneRoleResultById) == 1, Equals, true)
 
@@ -171,7 +175,7 @@ func (vcd *TestVCD) Test_OpenApiInlineStructCRUDRoles(check *C) {
 		check.Assert(err, IsNil)
 
 		oneRole := &InlineRoles{}
-		err = vcd.vdc.client.OpenApiGetItem(minimumRequiredApiVersion, singleRef, nil, oneRole)
+		err = vcd.vdc.client.OpenApiGetItem(apiVersion, singleRef, nil, oneRole)
 		check.Assert(err, IsNil)
 		check.Assert(oneRole, NotNil)
 
@@ -192,7 +196,7 @@ func (vcd *TestVCD) Test_OpenApiInlineStructCRUDRoles(check *C) {
 		ReadOnly:  false,
 	}
 	newRoleResponse := &InlineRoles{}
-	err = vcd.client.Client.OpenApiPostItem(minimumRequiredApiVersion, createUrl, nil, newRole, newRoleResponse)
+	err = vcd.client.Client.OpenApiPostItem(apiVersion, createUrl, nil, newRole, newRoleResponse)
 	check.Assert(err, IsNil)
 
 	// Ensure supplied and created structs differ only by ID
@@ -205,7 +209,7 @@ func (vcd *TestVCD) Test_OpenApiInlineStructCRUDRoles(check *C) {
 	check.Assert(err, IsNil)
 
 	updatedRoleResponse := &InlineRoles{}
-	err = vcd.client.Client.OpenApiPutItem(minimumRequiredApiVersion, updateUrl, nil, newRoleResponse, updatedRoleResponse)
+	err = vcd.client.Client.OpenApiPutItem(apiVersion, updateUrl, nil, newRoleResponse, updatedRoleResponse)
 	check.Assert(err, IsNil)
 
 	// Ensure supplied and response objects are identical (update worked)
@@ -215,19 +219,19 @@ func (vcd *TestVCD) Test_OpenApiInlineStructCRUDRoles(check *C) {
 	deleteUrlRef, err := vcd.client.Client.OpenApiBuildEndpoint(endpoint, newRoleResponse.ID)
 	check.Assert(err, IsNil)
 
-	err = vcd.client.Client.OpenApiDeleteItem(minimumRequiredApiVersion, deleteUrlRef, nil)
+	err = vcd.client.Client.OpenApiDeleteItem(apiVersion, deleteUrlRef, nil)
 	check.Assert(err, IsNil)
 
 	// Step 6 - try to read deleted role and expect error to contain 'ErrorEntityNotFound'
 	// Read is tricky - it throws an error ACCESS_TO_RESOURCE_IS_FORBIDDEN when the resource with ID does not
 	// exist therefore one cannot know what kind of error occurred.
 	lostRole := &InlineRoles{}
-	err = vcd.client.Client.OpenApiGetItem(minimumRequiredApiVersion, deleteUrlRef, nil, lostRole)
+	err = vcd.client.Client.OpenApiGetItem(apiVersion, deleteUrlRef, nil, lostRole)
 	check.Assert(ContainsNotFound(err), Equals, true)
 
 	// Step 7 - test synchronous POST and PUT functions (because Roles is a synchronous OpenAPI endpoint)
 	newRole.ID = "" // unset ID as it cannot be set for creation
-	err = vcd.client.Client.OpenApiPostItemSync(minimumRequiredApiVersion, createUrl, nil, newRole, newRoleResponse)
+	err = vcd.client.Client.OpenApiPostItemSync(apiVersion, createUrl, nil, newRole, newRoleResponse)
 	check.Assert(err, IsNil)
 
 	// Ensure supplied and created structs differ only by ID
@@ -240,7 +244,7 @@ func (vcd *TestVCD) Test_OpenApiInlineStructCRUDRoles(check *C) {
 	check.Assert(err, IsNil)
 
 	updatedRoleResponse2 := &InlineRoles{}
-	err = vcd.client.Client.OpenApiPutItem(minimumRequiredApiVersion, updateUrl2, nil, newRoleResponse, updatedRoleResponse2)
+	err = vcd.client.Client.OpenApiPutItem(apiVersion, updateUrl2, nil, newRoleResponse, updatedRoleResponse2)
 	check.Assert(err, IsNil)
 
 	// Ensure supplied and response objects are identical (update worked)
@@ -250,33 +254,19 @@ func (vcd *TestVCD) Test_OpenApiInlineStructCRUDRoles(check *C) {
 	deleteUrlRef2, err := vcd.client.Client.OpenApiBuildEndpoint(endpoint, newRoleResponse.ID)
 	check.Assert(err, IsNil)
 
-	err = vcd.client.Client.OpenApiDeleteItem(minimumRequiredApiVersion, deleteUrlRef2, nil)
+	err = vcd.client.Client.OpenApiDeleteItem(apiVersion, deleteUrlRef2, nil)
 	check.Assert(err, IsNil)
 
 }
 
-// skipOpenApiEndpointTest is a helper to skip tests for particular unsupported OpenAPI endpoints
-func skipOpenApiEndpointTest(vcd *TestVCD, check *C, endpoint, requiredVersion string) {
-	constraint := ">= " + requiredVersion
-	if !vcd.client.Client.APIVCDMaxVersionIs(constraint) {
-		maxSupportedVersion, err := vcd.client.Client.maxSupportedVersion()
-		if err != nil {
-			panic(fmt.Sprintf("Could not get maximum supported version: %s", err))
-		}
-		skipText := fmt.Sprintf("Skipping test because OpenAPI endpoint '%s' must satisfy API version constraint '%s'. Maximum supported version is %s",
-			endpoint, constraint, maxSupportedVersion)
-		check.Skip(skipText)
-	}
-}
-
 // getAuditTrailTimestampWithElements helps to pick good timestamp filter so that it doesn't take long time to retrieve
 // too many items
-func getAuditTrailTimestampWithElements(elementCount int, check *C, vcd *TestVCD, minimumRequiredApiVersion string, urlRef *url.URL) string {
+func getAuditTrailTimestampWithElements(elementCount int, check *C, vcd *TestVCD, apiVersion string, urlRef *url.URL) string {
 	client := vcd.client.Client
 	qp := url.Values{}
 	qp.Add("pageSize", "128")
 	qp.Add("sortDesc", "timestamp") // Need to get the newest
-	req := client.newOpenApiRequest(minimumRequiredApiVersion, qp, http.MethodGet, urlRef, nil)
+	req := client.newOpenApiRequest(apiVersion, qp, http.MethodGet, urlRef, nil)
 
 	resp, err := client.Http.Do(req)
 	check.Assert(err, IsNil)